# openapi/baseline.json
{
  "components": {
    "schemas": {
...
      "FileMetadataDto": {
        "properties": {
          "duration": {
            "format": "int64",
            "type": "integer"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "fileId": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "fileSize": {
            "format": "int64",
            "type": "integer"
          },
          "packetCount": {
            "format": "int32",
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "FileUploadResponse": {
        "properties": {
          "fileId": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "fileSize": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "storageLocation": {
            "type": "string"
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "MergeFilesRequest": {
        "properties": {
          "fileIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 2147483647,
            "minItems": 2,
            "type": "array"
          },
          "mergedFileName": {
            "type": "string"
          }
        },
        "required": [
          "fileIds"
        ],
        "type": "object"
      },
...
      "PagedResponseFileMetadataDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/FileMetadataDto"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "StreamingResponseBody": {
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/files": {
      "get": {
        "description": "Get all uploaded files with pagination",
        "operationId": "getAllFiles",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "uploadedAt,desc",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "enum": [
                "ANALYSIS",
                "MONITOR"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseFileMetadataDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get all files",
        "tags": [
          "File Management"
        ]
      },
      "post": {
        "description": "Upload a PCAP file for analysis",
        "operationId": "uploadFile",
        "parameters": [
          {
            "in": "query",
            "name": "enableNdpi",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "enableSuricata",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "enableFileExtraction",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "ANALYSIS",
              "enum": [
                "ANALYSIS",
                "MONITOR"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Upload PCAP file",
        "tags": [
          "File Management"
        ]
      }
    },
...
    "/api/v1/files/merge": {
      "post": {
        "description": "Merge two or more PCAP files into a single new file and trigger analysis",
        "operationId": "mergeFiles",
        "parameters": [
          {
            "in": "query",
            "name": "enableNdpi",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "enableSuricata",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "enableFileExtraction",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeFilesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Merge PCAP files",
        "tags": [
          "File Management"
        ]
      }
    },
    "/api/v1/files/{fileId}": {
      "delete": {
        "description": "Delete a PCAP file",
        "operationId": "deleteFile",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Delete file",
        "tags": [
          "File Management"
        ]
      },
      "get": {
        "description": "Get metadata for a specific file",
        "operationId": "getFileMetadata",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadataDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get file metadata",
        "tags": [
          "File Management"
        ]
      }
    },
    "/api/v1/files/{fileId}/download": {
      "get": {
        "description": "Download a PCAP file",
        "operationId": "downloadFile",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Download file",
        "tags": [
          "File Management"
        ]
      }
    },
...
    "/api/v1/files/{fileId}/extractions/{extractionId}/download": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "extractionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StreamingResponseBody"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Download an extracted file",
        "tags": [
          "Extracted Files"
        ]
      }
    },
...
    "/api/v1/signatures": {
      "get": {
        "operationId": "getSignatures",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get custom signatures file content",
        "tags": [
          "Signatures"
        ]
      },
      "put": {
        "operationId": "saveSignatures",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Save custom signatures file content",
        "tags": [
          "Signatures"
        ]
      }
    },
...
  },
...
}